\name{IntSequenceMemoizerParameters-class}
\alias{IntSequenceMemoizerParameters-class}
\alias{IntSequenceMemoizerParameters}
\alias{IntSequenceMemoizerParameters,integer,integer,integer,integer-method}
\alias{IntSequenceMemoizerParameters,integer,integer,missing,missing-method}
\alias{IntSequenceMemoizerParameters,integer,missing,missing,missing-method}


\title{IntSequenceMemoizerParameters}
\description{
  Wraps \code{edu.columbia.stat.wood.pub.sequencememoizer.IntSequenceMemoizerParameters}.
  This class holds the parameters used to create the SequenceMemoizer.
  The method for construction wraps those described in the javadoc  
}
\usage{
  IntSequenceMemoizerParameters (alphabetSize, depth, maxNumberRestaurants, maxSequenceLength, \dots)
}
\arguments{
  \item{alphabetSize}{See \url{http://www.sequencememoizer.com/documentation/sequencememoizer/index.html?edu/columbia/stat/wood/pub/sequencememoizer/IntSequenceMemoizerParameters.html}}
  \item{depth}{See \url{http://www.sequencememoizer.com/documentation/sequencememoizer/index.html?edu/columbia/stat/wood/pub/sequencememoizer/IntSequenceMemoizerParameters.html}}
  \item{maxNumberRestaurants}{See \url{http://www.sequencememoizer.com/documentation/sequencememoizer/index.html?edu/columbia/stat/wood/pub/sequencememoizer/IntSequenceMemoizerParameters.html}}
  \item{maxSequenceLength}{See \url{http://www.sequencememoizer.com/documentation/sequencememoizer/index.html?edu/columbia/stat/wood/pub/sequencememoizer/IntSequenceMemoizerParameters.html}}
  \item{\dots}{Not used.}
}
\references{
  \url{http://www.sequencememoizer.com/documentation/sequencememoizer/index.html?edu/columbia/stat/wood/pub/sequencememoizer/IntSequenceMemoizerParameters.html}
}

